+++ app/models/slot.rb
@@ -2,6 +2,7 @@
 
 # Time range of duration defined by ApplicationHelper::SLOT_DURATION, slicing an Availability.
 # During a slot a Reservation is possible
+# Only reserved slots are persisted in DB, others are instanciated on the fly
 class Slot < ActiveRecord::Base
   include NotifyWith::NotificationAttachedObject